requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.7.2
gTTS==2.5.1
edge-tts==6.1.6
